Move fetch to EventsView and wrap component in Suspense
