+++ apps/web/src/app/(default)/for-studenter/arrangementer/loading.tsx
@@ -1,0 +1,4 @@
+export default function Loading() {
+  // You can add any UI inside Loading, including a Skeleton.
+  return <p>Laster...</p>;
+}
